Move Result to class of its own
